pyproject.toml
[tool.poetry]
name = "osl-agent-prototype"
version = "0.1.0"
description = "Personal assistant agent prototype with OpenAI-backed planning and semantic memory."
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
openai = ">=1.43.0"
anthropic = {version = ">=0.34.0", optional = true}
google-generativeai = {version = ">=0.8.0", optional = true}
python-dotenv = ">=1.0.1"
playwright = ">=1.46.0"
appium-python-client = ">=3.2.1"
chromadb = ">=0.5.5"
python-arango = ">=8.2.5"
fastapi = ">=0.114.0"
uvicorn = {version = ">=0.30.0", extras = ["standard"]}
structlog = ">=24.4.0"
python-json-logger = ">=2.0.7"
cpms-client = ">=0.1.2"
sentence-transformers = "^5.2.0"
networkx = ">=3.2.0"
pillow = "^12.1.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.0"
pytest-asyncio = "^1.3.0"

[tool.poetry.scripts]
agent-service = "src.personal_assistant.service:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
